Change the fs_passno of nfs to 0
Set fs_passno, the sixth variable of fstab to "0" when the device is nfs.

The link of Bug 1186699:  bugzilla.redhat.com/show_bug.cgi?id=1186699
The bug is that kdump-initrd contains entry requesting nfs dump filesystem
to get filesystemchecked. And there is an erro message said that nfs need
be checked.

In this issue, there's no fsck for nfs utility, e.g fsck.nfs like other
file system. Whatever fs_passno 0 or 2 are passed, no fsck is executed
at all for nfs mount.But in dracut, set it to be 2 always, so the erro
message appear.

In the fstab,the sixth variable fs_passno stands for that the device need
checked or not,and dracut set it to "2".To fix this issue, it should
be "0" when the device is nfs.The third variable stands for the type of
the filesystem and we can use it to judge whether the device is nfs.
So when the third variable of fstab contains "nfs", the sixth variable
fs_passno should be set to "0".

Signed-off-by: Chao Fan <cfan@redhat.com>