modules: lvm: fix activationskip check
Fix boot breakage due to the activationskip check introduced in
cfa365a32d47 ("lvm: Don't activate LVs with activationskip set").

In 'man 8 lvs', it is stated that LVs with activationskip set, will have
a "k" flag in their lvs attribute listing.

The current implementation excludes LVs which do not have the "k"
flag set. This leads to the activation of LVs with the activationskip
set and the exclusion of LVs with no activationskip set: the check's
logic is inverted.

Fix this issue by skipping those LVs that have the "k" flag in their
attributes.

Signed-off-by: Nicolai Stange <nicstange@gmail.com>