Add new file to export the common functions
For now, it will raise many warnings and errors, if third part script
tries to source the dracut common script(eg. dracut-functions.sh).

There is a new script file dracut-export-functions.sh to be created
to solve this issue. This script file is independent with the dracut.
Any common function can be added in this file to be used by third part
script.

Signed-off-by: Minfei Huang <mhuang@redhat.com>