98integrity: Use /etc/ima as dir for IMA policy and config file
To sync with systemd, use the filepath /etc/ima/ima-policy as
the default file location for the IMA policy. At the same time we
move the ima config file location to /etc/ima/ima. Adapt the
documentation to the new path. Maintain backwards compatibility
by still reading the old files in case the new ones do not exist.

Signed-off-by: Stefan Berger <stefanb@linux.vnet.ibm.com>